feat: switch from modelscope to cnclip
