+++ server/clip_server/model/clip_model.py
@@ -1,7 +1,7 @@
 from clip_server.model.pretrained_models import (
     _OPENCLIP_MODELS,
     _MULTILINGUALCLIP_MODELS,
-    _DAMO_MODELS,
+    _CNCLIP_MODELS,
     _VISUAL_MODEL_IMAGE_SIZE,
 )
 
@@ -35,10 +35,10 @@ class CLIPModel(BaseCLIPModel):
                 from clip_server.model.mclip_model import MultilingualCLIPModel
 
                 instance = super().__new__(MultilingualCLIPModel)
-            elif name in _DAMO_MODELS:
-                from clip_server_model.damo_model import DamoModel
+            elif name in _CNCLIP_MODELS:
+                from clip_server.model.cnclip_model import CNClipModel
 
-                instance = super().__new__(DamoModel)
+                instance = super().__new__(CNClipModel)
             else:
                 raise ValueError(
                     'CLIP model {} not found; below is a list of all available models:\n{}'.format(
