+++ server/clip_server/model/cnclip_model.py
@@ -1,0 +1,39 @@
+# Originally from https://github.com/OFA-Sys/Chinese-CLIP. MIT License.
+
+from clip_server.model.clip_model import CLIPModel
+from clip_server.model.pretrained_models import _VISUAL_MODEL_IMAGE_SIZE
+from cn_clip.clip import load_from_name
+
+
+class CNClipModel(CLIPModel):
+    def __init__(
+        self,
+        name: str,
+        device: str = 'cpu',
+        jit: bool = False,
+        dtype: str = None,
+        **kwargs
+    ):
+        super().__init__()
+        self._name = name
+
+        self._model, self._preprocess = load_from_name(name, device=device)
+        self._model.eval()
+
+    @staticmethod
+    def get_model_name(name: str):
+        return name
+
+    def encode_text(self, input_ids: 'torch.Tensor', **kwargs):
+        return self._model.encode_text(input_ids)
+
+    def encode_image(self, pixel_values: 'torch.Tensor', **kwargs):
+        return self._model.encode_image(pixel_values)
+
+    @property
+    def model_name(self):
+        return self.__class__.get_model_name(self._name)
+
+    @property
+    def image_size(self):
+        return _VISUAL_MODEL_IMAGE_SIZE.get(self._name, None)
