+++ server/setup.py
@@ -48,7 +48,7 @@ setup(
         'jina>=3.12.0',
         'prometheus-client',
         'open_clip_torch>=2.8.0',
-        'modelscope',
+        'cn_clip',
     ],
     extras_require={
         'onnx': [
